edge_backend/requirements.txt
# == edge_backend/requirements.txt ==
flask
paho-mqtt
requests
openai
prophet
pandas